Crud and app update to better handle not found scenarios, document response codes.

+++ src/fastapi_social_network/crud.py
@@ -40,7 +40,8 @@ def create_user(db: Session, user: schemas.UserCreate):
 
 def get_posts(db: Session, skip: int = 0, limit: int = 100, user_id: int | None = None):
     if user_id is not None:
-        return db.query(models.Post).options(joinedload(models.Post.reactions)).filter(models.Post.owner_id == user_id).offset(skip).limit(limit).all()
+        return db.query(models.Post).options(joinedload(models.Post.reactions)).filter(models.Post.owner_id == user_id)\
+            .offset(skip).limit(limit).all()
     return db.query(models.Post).options(joinedload(models.Post.reactions)).offset(skip).limit(limit).all()
 
 
@@ -87,6 +88,8 @@ def delete_user_post(db: Session, post_id: int, user_id: int):
 # Reactions
 def react_user_post(db: Session, post_id: int, user_id: int, dislike: bool = False):
     db_post = get_post(db, post_id)
+    if not db_post:
+        return None
     if db_post.owner_id == user_id:
         raise NoPermission()
     reaction = db.query(models.PostReaction)\
@@ -100,30 +103,39 @@ def react_user_post(db: Session, post_id: int, user_id: int, dislike: bool = Fal
             user_id=user_id,
             dislike=dislike
         )
-    db.add(reaction)
-    db.commit()
-    db.refresh(reaction)
-    return reaction
+        db.add(reaction)
+        db.commit()
+    db.refresh(db_post)
+    return db_post
 
 
 def delete_reaction_user_post(db: Session, post_id: int, user_id: int):
+    post = db.query(models.Post).filter(models.Post.id == post_id).first()
+    if not post:
+        return None
     reaction = db.query(models.PostReaction)\
         .filter(models.PostReaction.post_id == post_id)\
         .filter(models.PostReaction.user_id == user_id).first()
     if reaction:
         db.delete(reaction)
         db.commit()
-        db.refresh(reaction)
-    return reaction.post
+        db.refresh(post)
+    return post
 
 
 def get_posts_reactions_by_user(db: Session, user_id: int, skip: int = 0, limit: int = 100, dislike: bool = False):
+    user = get_user(db=db, user_id=user_id)
+    if not user:
+        return None
     posts = db.query(models.Post).join(models.PostReaction).filter(models.PostReaction.user_id == user_id)\
         .filter(models.PostReaction.dislike == dislike).offset(skip).limit(limit).all()
     return posts
 
 
 def get_users_reactions_by_post(db: Session, post_id: int, skip: int = 0, limit: int = 100, dislike: bool = False):
+    post = get_post(db=db, post_id=post_id)
+    if not post:
+        return None
     users = db.query(models.User).join(models.PostReaction).filter(models.PostReaction.post_id == post_id)\
         .filter(models.PostReaction.dislike == dislike).offset(skip).limit(limit).all()
     return users
